Addressing ubuntu build failures
