+++ packages/zowe-explorer/__tests__/__integration__/bdd/step_definitions/config-editor/profileList.steps.ts
@@ -10,123 +10,10 @@
  */
 
 import { When, Then } from "@cucumber/cucumber";
-import { Workbench } from "wdio-vscode-service";
 import { expect } from "@wdio/globals";
+import { robustClick, dismissTutorialOverlay } from "./profileListHelpers";
 
 declare const browser: any;
-
-
-/**
- * Dismiss the tutorial overlay backdrop if it is currently blocking the UI.
- * The backdrop's onClick handler calls the tutorial's skip/close callback, so
- * a DOM-level click on it is all that is needed.
- */
-export async function dismissTutorialOverlay(): Promise<void> {
-    await browser.execute(() => {
-        const backdrop = document.querySelector(".tutorial-overlay-backdrop") as HTMLElement | null;
-        if (backdrop) {
-            backdrop.click();
-        }
-    });
-}
-
-/**
- * Robust click helper for webview elements in CI.
- * Dismisses the tutorial overlay if present, then uses DOM-level scrollIntoView
- * (avoids Actions API / CDP commands that fail with
- * "unknown command: Browser.getWindowForTarget") and retries both a native
- * WebDriver click and a DOM-level fallback up to `attempts` times.
- */
-export async function robustClick(element: any, attempts = 6, waitMsBetween = 300): Promise<void> {
-    await dismissTutorialOverlay();
-    await element.waitForExist({ timeout: 15000 });
-    await element.waitForDisplayed({ timeout: 15000 });
-    await browser.execute((el: HTMLElement) => el.scrollIntoView({ block: "center" }), element);
-
-    let lastError: any = null;
-    for (let i = 0; i < attempts; i++) {
-        try {
-            await element.click();
-            return;
-        } catch (err) {
-            lastError = err;
-            try {
-                await browser.execute((el: HTMLElement) => el.click(), element);
-                return;
-            } catch (err2) {
-                lastError = err2;
-            }
-        }
-        await browser.pause(waitMsBetween);
-    }
-    throw lastError;
-}
-
-
-export async function verifyProfiles(expectedTreeTitles: string[], expectedFlatTitles: string[], workbench: Workbench) {
-    const webview = (await workbench.getAllWebviews())[0];
-    await webview.wait();
-    await webview.open();
-
-    const appContainer = await browser.$("[data-testid='config-editor-app']");
-    await appContainer.waitForExist({ timeout: 10000 });
-
-    const profileList = await browser.$("[data-testid='profile-list']");
-    await profileList.waitForExist({ timeout: 10000 });
-
-    const viewMode = await profileList.getAttribute("data-view-mode");
-
-    await browser.waitUntil(
-        async () => {
-            const selector = viewMode === "tree" ? "[data-testid='profile-tree-node']" : "[data-testid='profile-list-item']";
-            const items = await browser.$$(selector);
-            return items.length > 0;
-        },
-        { timeout: 10000, timeoutMsg: "Profile elements not found within timeout" }
-    );
-
-    const selector = viewMode === "tree" ? "[data-testid='profile-tree-node']" : "[data-testid='profile-list-item']";
-    const elements = await browser.$$(selector);
-
-    const foundProfiles: string[] = [];
-    for (const el of elements) {
-        const profileName = await el.getAttribute("data-profile-name");
-        if (profileName) foundProfiles.push(profileName);
-    }
-
-    for (const title of expectedTreeTitles) {
-        expect(foundProfiles).toContain(title);
-    }
-    expect(foundProfiles.length).toBeGreaterThanOrEqual(expectedTreeTitles.length);
-
-    if (viewMode === "tree") {
-        const viewToggleButton = await browser.$("[data-testid='view-mode-toggle']");
-        if (await viewToggleButton.isExisting()) {
-            await robustClick(viewToggleButton);
-
-            await browser.waitUntil(
-                async () => {
-                    const updatedList = await browser.$("[data-testid='profile-list']");
-                    const updatedMode = await updatedList.getAttribute("data-view-mode");
-                    return updatedMode === "flat";
-                },
-                { timeout: 20000, timeoutMsg: "Failed to switch to flat view" }
-            );
-
-            const flatItems = await browser.$$("[data-testid='profile-list-item']");
-            const flatProfiles: string[] = [];
-            for (const item of flatItems) {
-                const profileName = await item.getAttribute("data-profile-name");
-                if (profileName) flatProfiles.push(profileName);
-            }
-
-            for (const title of expectedFlatTitles) {
-                expect(flatProfiles).toContain(title);
-            }
-            expect(flatProfiles.length).toBe(expectedFlatTitles.length);
-        }
-    }
-}
 
 Then("the profile tree should contain expected profiles from zowe.config.json", async function () {
     const appContainer = await browser.$("[data-testid='config-editor-app']");
